+++ src/lib.rs
@@ -60,9 +60,11 @@
 extern crate rustc_serialize;
 extern crate base32;
 extern crate crypto;
+extern crate rand;
 extern crate time;
 
 pub mod oath;
+pub mod pass;
 
 #[cfg(feature = "cbindings")]
 extern crate libc;