+++ tests/test_libreauth.c
@@ -45,6 +45,7 @@ int main(void) {
 
   nb_tests += test_hotp();
   nb_tests += test_totp();
+  nb_tests += test_pass();
 
   printf("Ran %d tests.\n", nb_tests);